delivery/requirements.txt
asgiref==3.6.0
asyncio==3.4.3
attrs==22.2.0
click==8.1.3
dnspython==2.3.0
Flask==2.2.2
graphql-core==3.2.3
greenlet==2.0.1
greenletio==0.9.0
h11==0.14.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
memphis-py==0.2.9
nats-py==2.2.0
protobuf==4.21.12
pymongo==4.3.3
pyrsistent==0.19.3
python-dotenv==0.21.1
uvicorn==0.20.0
Werkzeug==2.2.2
zipp==3.11.0
